% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brain2d.R
\name{ashape_polygon_sf}
\alias{ashape_polygon_sf}
\title{Convert a point cloud to polygonal geometry using an alpha hull}
\usage{
ashape_polygon_sf(pts, alpha = "auto", smoothing_factor = 3)
}
\arguments{
\item{pts}{Point geometry.}

\item{alpha}{Alpha radius or \code{"auto"}.}

\item{smoothing_factor}{Scaling applied when \code{alpha = "auto"}.}
}
\value{
An \code{sfc} geometry vector.
}
\description{
Convert a point cloud to polygonal geometry using an alpha hull
}
