% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brain2d.R
\name{auto_alpha}
\alias{auto_alpha}
\title{Estimate alpha radius for an alpha hull}
\usage{
auto_alpha(xy, k = 1, factor = 3)
}
\arguments{
\item{xy}{Matrix of scaled coordinates.}

\item{k}{Number of nearest neighbours.}

\item{factor}{Scaling factor for the median distance.}
}
\value{
Numeric alpha radius.
}
\description{
Estimate alpha radius for an alpha hull
}
