% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{clear_resource_cache}
\alias{clear_resource_cache}
\title{Clear the ggbrat resource cache}
\usage{
clear_resource_cache(cache_dir = ggbrat_cache_dir(create = FALSE))
}
\arguments{
\item{cache_dir}{Resource cache directory.}
}
\value{
The removed cache path, invisibly.
}
\description{
Clear the ggbrat resource cache
}
