% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{ggbrat_cache_dir}
\alias{ggbrat_cache_dir}
\title{Locate the ggbrat resource cache}
\usage{
ggbrat_cache_dir(create = TRUE)
}
\arguments{
\item{create}{Whether to create the directory if necessary.}
}
\value{
The normalized cache path.
}
\description{
Locate the ggbrat resource cache
}
