% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brain2d.R
\name{knn_density_filter}
\alias{knn_density_filter}
\title{Filter points by local kNN density}
\usage{
knn_density_filter(X, k = 10, keep_quantile = 0.5, dim = 2)
}
\arguments{
\item{X}{Coordinate matrix.}

\item{k}{Number of nearest neighbours.}

\item{keep_quantile}{Quantile of dense points to keep.}

\item{dim}{Point dimensionality.}
}
\value{
A list with filtered points and density diagnostics.
}
\description{
Filter points by local kNN density
}
